refactor client src files  to directories
