+++ client/src/header/ModalItem.jsx
@@ -1,6 +1,6 @@
 import { useNavigate } from "react-router-dom";
 import { useContext } from "react";
-import { UserContext } from "./App";
+import { UserContext } from "../App";
 import styled from "styled-components";
 
 const Div = styled.div`
