+++ client/src/login/Login.jsx
@@ -1,8 +1,8 @@
 import { useContext, useRef } from "react";
 import { Link } from "react-router-dom";
-import { getMe, loginUser } from "../services/userServices";
+import { getMe, loginUser } from "./services";
 import { useNavigate } from "react-router-dom";
-import { UserContext } from "./App";
+import { UserContext } from "../App";
 import Grid from "@mui/material/Grid";
 import TextField from "@mui/material/TextField";
 import Button from "@mui/material/Button";
